backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
google-genai>=0.8.0
web3==6.15.0
